ai/voice-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.0
python-dotenv==1.0.0
pydub==0.25.1
SpeechRecognition==3.10.0
httpx==0.25.1
python-multipart==0.0.6
